pyproject.toml
[project]
name = "p-sparsity"
version = "0.1.0"
description = "Reinforcement Learning for Learning Algebraic Multigrid Preconditioner Sparsity Patterns"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torch-geometric>=2.3.0",
    "pyamg>=5.0.0",
    "scipy>=1.10.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "tensorboard>=2.13.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.0.270",
    "mypy>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.0.270",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
